Return non-optional structs from Swift functions
Updated Swift templates to make struct initializers non-failable and to
return structs as non-optional from methods. This makes it consistent
with C++ where structs are returned by value and thus cannot be null.
This also simplifies using struct-returning methods in the calling
Swift code (i.e. by the user of the generated API).

Smoke tests are updated in a separate commit. Functional tests are
updated in another separate commit.

Change-Id: I8ad393bad81030396d94315e2bd0a83a89bbd4d6
Resolves: APIGEN-1326