Fix returning instances from Swift Listener callbacks
Updated swift/GetReference tempate to create a non-owning RefHolder
object and thus avoid a premature deletion of the real Swift object held
by RefHolder. This happens when an interface instance is returned from a
Swift callback that is called from C++.

Added a function regression test. Smoke tests are updated in a separate
commit.

Change-Id: I136aca795bd27d21bb7d217366137a46c8c0dc3e
Resolves: APIGEN-1429