Add infrastructure for catching fatalError() in Swift functional tests
Added infrastructure for catching fatalError() in Swift functional
tests. FatalErrorHelper class has to be injected into the framework that
is being created from the generated source, so that the default
implementation of fatalError() can be replaced with a custom one at
compile time.

A functional test utilizing FatalErrorHelper has to run its
code-under-test in a separate thread because fatalError() function never
returns (even if it's a custom implementation). This causes some memory
leaks, detected by valgrind. But there leaks are caused by test code
only, not the generated code, so they can be (knowingly) ignored.

Change-Id: Ieb223c93dcab44dd86d2cb643691011edcd632f0
Resolves: APIGEN-1427