Implement Listeners with return values
Even though the "Listeners with return values" for Swift was declared as
"implemented", it was actually rather incomplete. It worked only for
primitive types and instances.

Updated relevant Swift and CBridge template to add proper support for
the rest of the types. Also updated SwiftNameRules to use full names for
Dictionary references.

On Java side only ByteBuffer support was missing.

Change-Id: Iaa4f5a5f273915ef5117018e1e66723c195ee31b
Related: APIGEN-1434