Fixes failing tests and crashes in JNI code when host system is MacOS
There was wrong cast to void* to work with reverse caching.
Proxy cpp types are derived both from CppBaseProxy and from the type
which proxy should implemet. When elemets are being added to reverse
proxy then pointer to this last time is casted to void*.
Then when proxy is destroyed it should remove themself from reverse
proxy in its CppBaseProxy and CppBaseProxy* was casted to void*.
Sometimes these two pointers are the same, but on MacOS they are
different, so elements once added to reverse proxy were never removed
and getJavaObject was returning dead objects.

Relates to IOTSDK-1168

Change-Id: I537e50e5c3aa73fcca6b4f46f7e159db23828c3c