Unify Swift dictionary return conversion
To be able to use the same type based conversion, dictionaries must be
collected per type. Added a MapGenerator similar to how it's done for
Arrays.
Only conversion C->Swift is done so far, the other direction will
follow.
Changed array index behaviour slightly to be more uniform so the same
logic can be used in Swift.

Change-Id: I1235bcd3d2f5781796af977b9bd49d592956ed10
Relates-To: APIGEN-1453