Consolidate JNI conversion templates
Consolidated JNI conversion templates for instances, enums and structs
from multiple small files into a few bigger files. This simplifies the
maintenance of these templates. Enum and struct conversion templates
need to be extended for "Nullability" implementation (APIGEN-1454) and
this consolidation simplifies the process.

Also cleaned up the formatting for the affected templates (partially).

Smoke tests are updated in a separate commit. Functional tests are
unchanged.

Change-Id: Id340b350caab01fbc6e0a9eab852b888b6b45640