Refactor SwiftProperty to include accessors explicitly
This was a part of "noRead" implementation (APIGEN-1508). The
implementation itself was cancelled but the SwiftProperty refactoring is
useful enough to be salvaged. It simplifies both templates and unit
tests.

Also added a missing unit test for ExternalElementsValidatorPredicate
and fixed another unit test for it. This was also salvaged from the same
cancelled commit.

Change-Id: If9b4ce5b63e7e9276c1cc9d5dbd03b464286b3b4