Replace CollectionOf with Array in Swift
Replaced all usages of custom CollectionOf class in Swift with the
built-in Swift Array class. This improves the generated public API by
displaying a known built-in type (even though the Collection behavior
remains unchanged) and also avoiding having "where" constraints on
protocol associated type (which were ugly and cumbersome).

Removed the CollectionOf class itself as it became unused.

Smoke and functional tests are updated in a separate commit.

Change-Id: Id4a43fa352e8af32d2bf8ce6bed9725a763a346e
Resolves: APIGEN-1518