Added support to export library symbols.
Use -cppexport option to provide the export macro name. (e.g. MY_LIB) This
will be used to generate macros for exporting the library symbols. When no
-cppexport option is used, _GENIUM_CPP will be implicitly used. The file
Export.h is generated and uses the following macros:
- MY_LIB_EXPORT: the macro defined by the header to perform the actual
  export.
- MY_LIB_SHARED: the macro that should be defined when building as a shared
  library.
- MY_LIB_INTERNAL: the macro that should be defined when building the
  library for generated code.

This is useful for two situations:
- Building as a DLL on the Windows platform, which requires explicit symbol
  exporting.
- Setting symbol visibility to hidden by default on UNIX platforms. This is
  useful to allow the linker to produce more optimal libraries or to
  enforce using only using public symbols.

A pre-defined export (_GENIUM_C_EXPORT) is used for cbridge code to always
export symbols.

Change-Id: I19317c58562ca47bcd9c6325c296bb3efc06861c
Resolves: APIGEN-1490