Added support to export library symbols.
Use -cppexport option to provide the export macro name. (e.g. MY_LIB) This
will be used to generate macros for exporting the library symbols. The file
Export.h is generated and uses the following macros:
- MY_LIB_EXPORT: the macro defined by the header to perform the actual
  export.
- MY_LIB_SHARED: the macro that should be defined when building as a shared
  library.
- MY_LIB_INTERNAL: the macro that should be defined when building the
  library for generated code.

This is useful for two situations:
- Building as a DLL on the Windows platform, which requires explicit symobl
  exporting.
- Setting symbol visibility to hidden by default on UNIX platforms. This is
  useful to allow the linker to produce more optimal libraries or to
  enforce using only using public symbols.

The examples now export symbols explicitly and set the default visibility
to hidden to verify the functionality.

Signed-off-by: Aaron Barany <aaron.barany@here.com>