Fixes for crash on exit when running swift example with shared libs.
A global static variable for a shared_ptr to a proxy object was used. Since
a global mapping of these objects is maintained, this would crash if the
tables were destructed before the shared_ptr was destructed. Two fixes are
done for this:
1. A function static variable is used for the cache maps. These will be
   constructed on first use, meaning that static initialization that
   constructs a proxy object should work properly. C++11 guarantees that
   this initialization is thread-safe.
2. An "alive" flag is used, which is cleared out on destruction. This
   protects against a proxy object being destructed after the static map is
   destroyed. For example, if an instance is stored in a global variable
   and destructed after the cache maps.

To add more to the safety, operator[] is no longer used on the map on
destruction. This will insert an item if it's not already present, and is
also inefficient on some implementations. (it will allocate a node even if
it's already present, then remove it afterward) It additionally avoids a
second key lookup for the erase.

Change-Id: If8ef5f76ddac390989d2b44ebd17429e30e38377
Relates-To: APIGEN-1490