Add dedicated JNI field getters for unsigned ints
Updated JNI templates to use "overloads with a dummy pointer" type
resolution technique for field getter functions (similarly to what is
already done for convert_from_jni() family of functions). This allows
for dedicated functions for unsigned int types, even though the
underlying Java type is the same as for signed int types.

This is a part of the fix for "nullable unsigned ints" compilation
issue.

Smoke tests are updated in a separate commit. Functional tests need no
updates, as the change is a refactoring for all types but unsigned ints.
Functional tests for unsigned ints will be added later when setter
functions are fixed as well.

Change-Id: I0591f6273f9620fe7014f11c9b663d6af97ea8d2
Related: APIGEN-1538