Fix compilation issues for Struct with Immutable fields
Updated JNI and CBridge templates to create C++ structs in one
constructor call if one of the fields has an immutable struct type. Same
approach was already implemented for structs that are immutable
themselves. This solves compilation issues in JNI and CBridge.

Added compilation functional test. Smoke tests are added/updated in a
separate commit.

Change-Id: Ideaae3ac2878b81193d9fc2396646fdec170804d
Resolves: APIGEN-1546