Use type-aliased name for Maps/Dictionaries in Swift
Updated Map/Dictionary templates in Swift to use a type-aliased "public
name" for the generated code, instead of a raw [key:value] dictionary
definition. This looks much more readable to the user of the generated
API.

Also removed SwiftType.implementingClass field as it became redundant.

Updated smoke tests. Functional tests were not updated as this change
does not affect compilation of the generated code.

Change-Id: Id3d20538037b0fb6be411509ab20ea5d66e5de01
Related: APIGEN-1536