Add optional implementation with unit tests
Optional.h header is supposed to be used throughout Genium and is used
to switch between different optional implementations. If nothing is
explicitly defined it tries to detect whether or not std::optional is
available and then uses it or falls back to our own implementation.
Our own fallback should behave mostly like std::optional.

Change-Id: I2a816f8ea91471641828b2c469871404ea4e6104
Relates-To: APIGEN-1552