Improve DeclarationOrder smoke test
Reordered type declarations in DeclarationOrder.fidl so that they are
guaranteed to be in the wrong order in C++ generated code if
TopologicalSort algorithm is fully or partially unapplied.

Also updated the Map type definition in DeclarationOrder.fidl to have an
Array value type so that the mutual ordering of those two is tested as
well.

Change-Id: I8e552ba3ad3eeb350f323ca995e8f736f2b82d02