Add GTest tests for cpp generator
These tests replace the example C++ app since it did not showcase much.
GTest is added as external project and downloaded from git. Since it
needs to be compiled with the same compiler options there is not
prebuilt binary available.
Added the ability to compile the test binary with different C++
standards. The build-cpp scripts installs these into different folders
so it's easier to run on CI.

Relates-To: APIGEN-1566
Change-Id: Ibbcf6b8e39dd70e80bc1fe586e5305accb37784c