Hide internal classes and constructors from doc tools
Updated the templates to mark "internal" types and constructors with
special markers to make documentation tool skip those ("@exclude" for
Java, ":nodoc" for Swift).

Also updated the copyright header for NativeBase.java to match the
copyright header of other "static" files (i.e. files where the code
content is static and only package/namespace is inserted in a template).

Change-Id: I38f02935aa10d6e6d961862f1acc92f7bd078d9b
Resolves: APIGEN-1579