Prefix accessor method comments with verbs in Java and C++
Updated Java and C++ model builders to pre-process doc comments for
attribute accessor methods by prefixing those with an appropriate verb
(and adjusting the sentence capitalization accordingly).

Swift generated code does not require this preprocessing as it
represents FIDL attributes as Swift properties, not methods.

Also documented the documentation conventions in a new .md file.

Smoke tests are updated in a separate commit.

Change-Id: I87aa222f9562943278e89857019e7609dfeebf16
Resolves: APIGEN-1587