Make output parameter nonmandatory for constructors
Updated model builders to infer the return type for "Constructor"
methods automatically instead of relying on FIDL definition. This way
the return type declaration for constructors in FIDL is nonmandatory. It
is still needed if there is a documentation comment for the return type,
but can be safely omitted otherwise.

Change-Id: I722d3495c95e526407a80f353ad4bdcfaa8f3441
Resolves: APIGEN-1589