Treat arrays of immutable structs as immutable in C++
Updated C++ model builder to mark fields as immutable if the field type
is an array of immutable structs. This fixes CBridge and JNI compilation
issues for this scenario.

The new implementation also handles typedefs to immutable structs and
maps with immutable structs as a value type (structs are not supported
as a map key type).

Added functional compilation test as a regression test.

Change-Id: Id041087dc2db292f35886541962a0c17dc9eb55b
Resolves: APIGEN-1600