Add FDEPL support for structs with methods
Added a new "Companion" FDEPL property for structs. This property should
contain a fully-qualified FIDL name of an interface that defines methods
for this struct. See "franca.md" documentation for details.

This commit contains only Franca-to-LIME implementation part.
Implementation for target languages is added in the follow-up commits.

Added FIDL/FDEPL files for smoke and functional tests.

Change-Id: Iaed76ba4af011870336fbc3557ca54f634aeb57e
Related: APIGEN-1511