Generate comment links for Swift
To be able to collect all model types and resolve all links
SwiftGenerator was changed to return the SwiftFile as model instead of
already generated file. This made it necessary for SwiftFile to know its
file name. In order to resolve the names easily the Swift model was made
streamable just like the Java model.

Change-Id: I95bda44b33e43bf2deea667ca8a01d08b959b910
Relates-To: APIGEN-1591