Refactor LIME type references
Extracted LimeTypeRef interface from the class previously known as
LimeTypeRef, to enable having more different type reference kinds. A new
reference kind would be needed for the upcoming custom IDL, to handle
type refs in presence of "import" statements.

Renamed LimeTypeRef class to LimeLazyTypeRef, to better indicate that
it's the only type reference kind to exhibit the "lazy" behavior.

Introduced LimeDirectTypeRef for the cases when the exact type is known
immediately and the lazy behavior is not required.

Change-Id: I836c55d8dc01f7e5dd8252f6a63ebfa19f448eb3
Related: APIGEN-1578