Fix "listeners with exceptions" support
Updated "java/ExceptionDefinition" template to expose the constructor
for the exception class as "public". This way the user of the API could
properly implement a throwing "listener" interface. Updated functional
tests to include this scenario.

Also updated CBridgeComponents includes look-up to fix a missing CBridge
include for the same scenario.

Change-Id: I13881f2e8dcbd53434015c4ae198c60957c6986e
Resolves: APIGEN-1624