Add helpers to allow equatable classes as keys in sets/maps
Often users want to use equatable and hashable types as keys in maps or
sets. As complex types are always handled with std::shared_ptr this
should be used in these maps as well. The member operator== of
std::shared_ptr however only compares the pointer value. To be able to
use equatable classes as keys, helper classes EquatableHash and
EquatableEqualTo are generated.

Change-Id: I0eaea1fff7928b003768162c1388d339f01d35f3
Relates-To: APIGEN-1635, APIGEN-1488