Remove PMD and Checkstyle linters
PMD and Checkstyle linter tools were necessary in the early days of
Genium code base, when the main language of the code base was Java.

Currently the existing code base is being gradually migrated to Kotlin
and all new code files added are exclusively in Kotlin as well. This
diminishes the value of these linter tools considerably, as they do not
lint Kotlin files.

Moreover, these linters add to build time and least on of those linters
have dependencies with problematic licenses, hindering the open-sourcing
process.

Considering all of the above, removed PMD and Checkstyle linters from
Genium gradle scripts.

Change-Id: Idabc50eb8bc78058cab470972ea3a071e5a2e338