Add LimeException class to LIME model
Added LimeException class to LIME model to remove direct coupling
between methods and error enums. The direct usage of enums as error
types is Franca-specific. The new custom IDL (APIGEN-1578) needs to
improve the declaration of exceptions: the need names, attributes and
documentation of their own. With this decoupling such declaration will
be possible.

Change-Id: I8583ef53a45d5777543692b6957589898af2eeaa