Change "Constructor" LIME attribute into a Boolean flag
Updated LIME model to represent constructor methods with an
"isConstructor" flag on LimeMethod instead of a "Constructor" attribute
on LimeMethod. The intent is to align the internal representation to the
planned syntax in the new IDL language (planned syntax is "constructor
foo(...)" instead of "@Constructor function foo(...)" as it would be
with an attribute).

Change-Id: I67b1579a65e464010df706cf0d18d1df479c3b51
Related: APIGEN-1578