Sort JNI includes and elements
Updated JNI generator to explicitly sort lists of includes and lists of
elements just before applying Mustache template to those lists. This
makes the generated code stable with regard to declaration order and
with regard to file loading order.

In addition to just making JNI output more stable, this is also required
for transition to the new IDL language (APIGEN-1671), as the declaration
order of the new input files might differ from the order in the old FIDL
files.

Change-Id: Ib385233bc1da8fbe133919d615a84422cad96554