Add toString() implementation to LIME enums and values
Updated most enums in the LIME model class hierarchy, as well as
LimeAttributes and LimeValue classes, with custom toString()
implementations, as needed for the upcoming LIME generator.

Also updated LimeArray, LimeMap and LimeSet with meaningful
implementations for "name" property. This is needed for the LIME
generator too, but also improves Genium debug output.

Change-Id: I194da3d0dddea03faebeafa49f68c8eac364ff9a
Related: APIGEN-1671