Refactor handling of String and Special literals in LIME model
Updated model builders for both LIME and the output languages to escape
string literals on output through LimeValue.Literal.toString() instead
of escaping them on input.

This simplifies handling of literal constants in LIME model in general,
not just string literal, because now there's only a single place where
the string literals need to be handled in a special way.

This simplification is needed for LimeIDL parser implementation, to
avoid adding this special handling to yet another place.

Also updated the type checks that are performed for LimeValue.Special
literals to avoid reference equality check and check for
LimeBasicType.TypeId instead.

Change-Id: Ibd02ce2912b02a0061ee327c79a290120613724b
Related: APIGEN-1672