Improve handling of method overloads
Updated model builders to detect whether a method is overloaded based on
the methods signature, instead of just treating every method with a
disambiguation suffix (aka "selector" in Franca) as overloaded.

This is needed for LimeIDL loader implementation, as LimeIDL does not
have a concept of a user-defined disambiguation suffix.

Change-Id: I80d7b9e35a11032d811a3f19fb4d191e0f9a2701
Related: APIGEN-1672