Refactor LimeContainer class, part 1
Refactored LimeContainer class into separate classes for "class",
"interface" and "types" elements. In its previous form it was a remnant
from Franca times and thus was very confusing, even hard to document
properly.

This commit contains the changes for "lime-runtime" and "lime-loader"
submodules. "genium" submodule is covered in a separate commit.

Change-Id: I78fd910ba0f38a4e6e901bc083379434eb24af61
Related: APIGEN-1582