Add static checks for "external" types in C++
Updated C++ Mustache templates for "external" types to include
static_assert() checks that make sure the actual "external" types
matches its IDL definition.

For a class, the return types of the methods are checked. This has a
side effect of failing the compilation if the function names or function
parameter types do not match.

For a struct, the types of fields are checked. This has a side effect
of failing the compilation if the field names do not match. For fields
that have "external" accessors defined, method checks are done for the
accessors instead.

For an enum, enum values are checked (both explicit and implicit
values are checked). This has a side effect of failing the compilation
if the enum value names do not match.

Change-Id: If9a188ac1d948c08ce623b1cc904344915ccc8a1
Resolves: APIGEN-1593