Preserve subclass type from C++ to Swift
CBridge code uses C++ generated stored type information to look up the
correct type of the class. Bridging code has its own table where the
mapping of type id to Swift init function is stored. This table is
global as it's only initialized once statically and afterwards only read.
In order for C++ code being able to call into Swift, the construction
method has to be annotated with @_cdecl. Although this is internal, it
exists since Swift 3 and is used in Swifts code as well, so it's likely
going to stay.

Change-Id: Id25af80972213d5aabf1ad802c422c436f1f424a
Relates-To: APIGEN-1618