Remove cache unit tests
Removed unit tests for the output file cache mechanism. These tests were
having the file system mocked away, which is correct for unit tests but
is rather useless for testing a cache mechanism. File system is the main
point of failure for an on-disk cache.

These tests need to be replaced by a set of functional tests. Meanwhile,
it is preferable to remove them as they actually don't test much while
preventing any refactorings of the cache code due to some intrusive
mocking involved.

Change-Id: I0ec05626a9f35f7c159b70b3fd3157dd3e652ce5
Related: APIGEN-1641