Add @Cpp(Accessors) attribute to LimeIDL
Added @Cpp(Accessors) attribute to LimeIDL and added support for this
attribute to C++ generator. A struct marked with this attribute will
have it's field generated as "private" in C++ and with have "public"
accessor methods generated for the fields.

When used in combination with @Immutable attribute, only getter methods
are generated. Moreover, with @Immutable the private fields are not
"const", thus allowing the C++ compiler to produce implicit copy
constructors for the struct type. This enables the usage of such structs
in std::vector<>, among other things.

Added compilation functional tests for C++ and glue layer code. Added a
separate compilation test for checking copy construct-ability in C++.

Change-Id: Ifd83a923178232d89d3443fbba7a6d7c50564aa7
Related: APIGEN-1520