Update Java model to store all class names
Updated Java model to store all class names for a nested types (i.e. the
names of the types nesting "parents").

Updated JNI model to store a concatenation of these class names as a JNI
class name, instead of trying to figure out nesting dynamically. This
concatenated name is the only Java name JNI needs. Single name of just
the type itself is not useful anywhere in JNI.

This update to the naming logic is needed to support nested classes
implementation (APIGEN-1727).

With regard to generated code, this change is a refactoring, so no smoke
tests were updated.

Change-Id: I249d031ef5ac1a186776f332ad4852c9385719bd
Related: APIGEN-1727