Don't walk inheritance parent in LIME tree
Updated LimeTreeWalker to stop walking inheritance parent as a type.
Some model builders require full information about the parent type, but
the type itself should not end up in the final results and thus is
transient.

Updated model builders that require extended inheritance parent
information to build that parent as a transient model on their own. The
process of building a transient model is abstracted away behind a simple
lambda. This keeps the coupling to the actual code behind the lambda
loose and enables easy faking for unit tests.

Change-Id: Icb6dbb19fc38a23ec332dd4ae977ee6a555ee3f9
Related: APIGEN-1727