Remove offending XCode environment variable
XCode 11 adds an environment variable to invocations of external scripts
which causes Swift compiler to look for a `.SwiftFileList` file. It also
leads to framework `.swift` files being picked up by XCode for the
application target, which shouldn't be the case and fails because of
duplicate file names. Unsetting this in the build script restores pre
XCode 11 behaviour. Also moved the build directory out of the test
project to be sure the sources are not picked up by XCode.

Change-Id: Ib410a83b3356230ad9243d55358450e42834548b
Relates-To: APIGEN-1742