Skip includes for forward-declared types in C++
Updated C++ generator to skip includes for forward-declared types in
generated header files. Includes are still added for corresponding
implementation files. This fixes an issue of failing compilation for a
circular dependency between types when one of the types is a nested
type.

Deleted CppTemplateTypeRefTest tests because they were not testing  any
logic.

Smoke and functional tests are updated in a separate commits.

Change-Id: I65636584e404c3033ba5bf16d1e5693645549692
Resolves: APIGEN-1746