Allow any type as exception payload in C++
Updated C++ model builder to allow any type as an error type, instead of
allowing only enums as before.

Added a Return<void, Error> specialization to Return<Value, Error> C++
template. This allows for a consistent C++ return value type for methods
with errors, regardless of whether the IDL return type is "Void" or not.

Smoke and functional tests are added in a separate commit.

Change-Id: I44585098c2abe074f04066d2c7fe271079dc705a
Related: APIGEN-1747