Fix lambdas in Swift generator
Updated SwiftGenerator and SwiftGenerator classes to fix the issue where
a generated file that consisted of a single lambda declaration was not
committed to disk.

Updated handling of nullable type references in SwiftModelBuilder and
CBridgeModelBuilder to be done in their respective type mappers instead.
Limiting this handling to a single place ensures there is no partial
loss of nullability information (which was an issue before for lambdas).

Added/updated smoke and functional tests accordingly.

Change-Id: I6c4dd951f46a2b1d24db4fa975a20f43dabc4233
Resolves: APIGEN-1751