Add regression test for missing self-include
Added a functional compilation test as regression test for missing
"external" header include. The issue happens when the package names and
the class name for the IDL class are the same as directory names and
file name for the .h include.

The issue was fixed some releases ago as a side effect of some other
change, but it had no dedicated regression test until now.

Change-Id: Iabd199ac6eb59f85f8998261d99d7da9993ce8ad
Resolves: APIGEN-1754