Clean up genium-cmake scripts for Swift
Main objective was to make apigen_swift_test independent of Genium
generation, i.e. it should be usable with non-generated targets. This
will be used for future tests with multiple generated targets.
This is mostly related to Linux, since we use normal binary targets on
mac OS and CMake is doing the bulk of the work there.
To be able to make apigen_swift_test independent from
apigen_swift_compile all necessary flags to link the generated Swift
module from the tests are added to the compiled target and propagated as
INTERFACE_LINK_LIBRARIES, INTERFACE_LINK_DIRECTORIES and
INTERFACE_INCLUDE_DIRECTORIES.

Change-Id: Iabedab7f5856b10580db3bf641e9936348bb7764