Don't use reference map for error enum analysis in C++ generator
Updated BaseApiGeneratorSuite to stop using referenceMap.values as a
"convenient" flattened representation of LIME model. With the
introduction of auxiliary sources, the reference map can now be wider
than the actual non-flattened model (limeModel.topElements).

While this particular code in BaseApiGeneratorSuite does not produce an
incorrect result with auxiliary sources, it represents a bad practice
that could be error-prone if employed elsewhere.

Change-Id: I6c5cfaf14993327c50ae2db77f2519c0e5cfa3da
Related: APIGEN-1756