Make common output directory configurable for C++ and Java
GeneratedFiles have now a source set (common or main) assigned by the
generators and Genium will output the files in the configured folders.
To allow nesting of the common and main directory, an adapter class is
introduced which takes care of using one cache for nested or two caches
for distinct folders.

Change-Id: I7cab93b6daca8aa5bfdb3fad123c4494cae9d264
Relates-To: APIGEN-1700