Fix platform names for Lambdas in Swift
As CBridge functions are not public, it does not matter whether their
names conform to @Swift(Name) attribute or not. Updated
SwiftModelBuilder class to ignore platform names for calling CBridge
bindings for Lambda, making this consistent with what is already
generated for CBridge.

This fixes a Swift compilation issue where a Lambda with a @Swift(Name)
platform name failed to compile, due to discrepancy in naming in Swift
code and CBridge code.

Added smoke test. Added functional compilation test as a regression
test.

Change-Id: I2192910cb20013fda2213b5481aee9442b890bc8
Resolves: APIGEN-1774