Add support for structs to Dart generator
Updated Dart generator to generate Dart and FFI files for converting
structs between Dart and C++ in both directions.

Added Dart and FFI Mustache templates for this conversion.

Updated dart/DartClass template to account for the fact that Dart
language does not allow nested types.

Updated dart/DartStruct template to provide an all-fields constructor
for the struct.

Smoke and functional tests are added/updated in a separate commit.

Change-Id: Ib0ab616218deeeb0993054516b64df230dd116a5
Resolves: APIGEN-1777