Fix JNI conversion include order dependency
JNI array conversion code relies on all conversion overloads being
visible. This can be easily achieved if all specific headers are
included first before the common headers.
For the case where external genium generated is used, the corresponding
headers need to be included first as well. Simplest solution for this is
to include all dependencies in the conversion headers rather than just
in the implementation file.

Signed-off-by: Georg Bremer <georg.bremer@here.com>