Update Robolectric to version 3.7.1
Robolectric 3.4.2 started to fail during its Maven dependency resolution
due to Maven Central HTTP access being shut down on 15.01.2020.
Upgrading to the latest 4.x version is infeasible as there are API
incompatibilities. Upgrading to 3.7.1 fixed the HTTP issue.

The HTTP issue is reproducible locally as well as on the legacy CI. Does
not seem to affect Travis CI at the moment, but this might be due to
some additional caching done there.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>