Add support for methods in structs to Dart generator
Moved generated struct conversion function from a separate Dart file
into the same where Dart struct definition resides. This is required
since the all-fields constructor for a struct is "private" if custom
constructors are defined. As there is no "internal" visibility in Dart
language, the only way to access the "private" (actually file-private)
constructor is to put everything into the same file.

Updated Dart templates to support methods (and constructors) on structs.

Added/updated smoke and functional tests for both methods and constants
on structs.

Resolves: #65
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>