Add support for "internal" visibility to Dart generator (#133)
There is no "internal" visibility concept in Dart. The solution is to
mark "internal" members (functions, properties, constants, fields) as
"private" with "_" prefix; and to mark "internal" types as "public", but
then to omit those types from the list of library exports.

"_" members marking was already in place. Updated Dart templates to not
to apply this marking to types. Implemented exports omission by
filtering the exportsCollector list in DartGeneratorSuite.

Added appropriate smoke tests.

Resolves: #130
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>