Remove custom ios toolchain (#144)
Removes custom toolchain to use default CMake's toolchain

Custom toolchains provide different options, variables and configurations.
They are not standardised and vary from project to project. It's hard to rely on
them  and it's not possible to mix toolchains. If there is existing project with its
own toolchain tries to integrate Gluecodium it might be problematic to sync
and adopt toolchains.

CMake 3.14.x introduced its own 'toolchain' which provides well documented
variables and it's possible to mix it with custom toolchain.