Add optional prefix to collection conversion functions in Swift
Updated Swift templates to add an optional prefix to the names of
collection conversion functions. This is needed for modularization
support (#167). CBridge part will be done separately.

Added a command line parameter for this prefix and CMake support for
this command line parameter.

Added "libhello_" Swift prefix to functional tests to test this
functionality.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>