Clean up "token" cache in Dart
Dart objects passed down to C++ are cached on Dart side int
"token-to-object" cache. If no cleanup is done for this cache, the
objects there are "leaked", never being garbage-collected.

To fix this, updated Dart and FFI templates to have the destructor of
the C++ proxy object clear the associated token-object pair from the
cache.

Resolves: #166
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>