Improve generated documentation for @FunctionalInterface in Java
Updated LimeLambda and LimeLambdaParameter classes to allow for
representing dedicated comments on lambda's parameters and return type.

Updated JavaModelBuilder to propagate these dedicated comments into a
structured JavaDoc comment on the method of the lambdas's
@FunctionalInterface.

Updated "Comments" smoke test to include a use case for structured
comments on a lambda.

Resolves: #37
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>