Fix conversion functions for Swift collection types
Updated SwiftTypeMapper to accumulate collection types only when
mapType() is called on such type directly, but not for other cases (e.g.
when a collection type is a parameter or a return type of a lambda).

Previously, over-eager accumulation of generics by SwiftTypeMapper led
to extraneous Swift conversion functions for collection types in
modularized builds. This failed compilation, as corresponding CBridge
functions were not present (because CBridge accumulation logic was
correct).

Added a unit test as a regression test.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>