Add support for "default" (nameless) constructors in Dart.
Added parser support and template support for marking class and struct
constructors as @Dart(Default). This forces the marked constructor to be
generated as a nameless constructor in Dart. This usage is only
meaningful in case of several overloaded constructors: if there's only
one constructor it will be generated nameless by default.

Added smoke and functional tests as appropriate.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>