Add support for cross-thread callbacks in Dart
Updated Dart and FFI template to add support for cross-thread callbacks
in Dart. Dart callbacks could only be called from within the same Dart
isolate they originate from. To facilitate this, a marshalling mechanism
was added that schedules the callbacks to be run on the correct isolate.

Added functional tests. Updated smoke tests.

Resolves: #138
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>