Add support for C strings in C++ generated code
Added new LimeIDL attribute @Cpp(CString). This attribute marks a String
type reference in IDL to be treated in C++ as const char* instead
of std::string. It has no effect on other types.

Updated C++, Swift and Java generators and templates to support this.

Added smoke and functional tests.

Resolves: #268
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>