Fix Dart conversion issue for enums with explicit and implicit values
Updated LimeEnumerator class to carry both explicit value (IDL
user-assigned) and computed value (previous enumerator value + 1, or 0
for the first enumerator).

Updated Dart template for enums to use computed value if explicit value
is not specified. This fixes a run-time conversion issue for enums with
a mix of explicit and implicit enumerator values.

Added smoke and functional tests.

Resolves: #279
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>