Fix ref equality for Java constructors (#312)
JNI implementation for Java constructor is not able to use "wrapper
cache" to cache the object being create (because the object does not
exist yet). This leads to corner-case bugs where a constructor-created
object is breaking referential equality on Java side.

Updated Java and JNI templates to circumvent this. Added functional and
smoke tests.

Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>