Improve documentation comments in generated Dart code (#340)
Updated LIME loader to correctly fill in path information for all
LimeComment comments.

Updated DartNameResolver to put the first sentence of a doc comment into
a separate paragraph, but only if it's a comment to a type or to a
function. This makes it more compliant with Dart documentation
guidelines.

Updated Dart template for function doc comment to look less like JavaDoc
and more like plain text. This makes it more compliant with Dart
documentation guidelines.

Resolves: #339
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>