Move most validators from lime-loader subproject to main subproject
Moved all validators except LimeImportsValidator from "lime-loader"
Gradle subproject to "gluecodium" Gradle subproject. The moved
validators validate LIME model and are agnostic to the source of that
model. Moving them to the main subproject improves separation of
concerns between subprojects. This also improves the support for
(currently unused) feature of using several different loaders for the
LIME model.

LimeImportsValidator however is validating imports, which is a feature
of LimeIDL language, but not LIME model. This validation is therefore
specific to the loader and should stay there.

Moved appropriate unit tests as well.

Resolves: #329
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>