Update C++ generator to use LIME model directly
Updated C++ generator classes and templates to use LIME model directly
instead of using an intermediate C++-specific model tree. Unfortunately,
some of the "old" infrastructure (CppModelBuilder and related classes)
cannot be deleted immediately, as other generators use those as well.
Due to this,  some of the classes of the "new" infrastructure have names
prefixed with "Cpp2". This is a temporary measure.

There are also minor changes in smoke tests. Most of those are related
to includes that were missing or extraneous in the old implementation.

Resolves: #353
Signed-off-by: Daniel Kamkha <daniel.kamkha@here.com>